requirements.txt
flask
flask_sqlalchemy
pytesseract
pillow
pdf2image
